SAK-32410 Additional changes and bug fixes for NYU Skin contributions (#4201)
* SAK-32410 Additional changes and bug fixes for NYU Skin contributions

Header Row Updates
* move the role switch action into a drop down menu
* only show role switch menu option if it has had the javascript set it up
* show role options on dekstop/tablet and offer menu on phone
* show the exit roleSwitch anchor in the navbar on all viewports

Updates to "Sites" modal
* Add NYU styles for all sites buttons and menus
* fix auto favorite panel on mobile
* make mask white
* update all term-level star size and make more clickable. Also show the close button when on mobile

Left Menu Updates
* Fix the toolmenu bg color and change blue color
* Tinker styles to extend the toolmenu to the top of its container, to remove the grey bit at the top

Update to Tool Content Area
* Apply NYU tool tab styles
* Align tabs with second tool menu item where possible
* Align the breadcrumb with the toolmenu's first item
* Apply NYU button styles
* Disable swapped-view styles by default (can be re-enabled via a skin variable)
* Apply NYU button styling to synoptic tool buttons
* Update the breadcrumb link color

Mobile Updates
* Apply NYU mobile design to Sakai 12
* Make the skipnav actions into two rows, one for the site-context and the other for the tool-context
* Don't uppercase the skipnav text and adjust the breadcrumb widths so long site and tool names clip nicely
* Affix only the banner on mobile and allow click to return to top
* Fix javascript lessonbuilder due to fixed header changes
* Fix button container view positioning in single column on mobile
* Fix help buttons and button container positioning on mobile

Favorites row
* Update the selected site's bg color

Lessons toolbar and navigation changes
* prevent lessons tabs and buttons from overlapping one another and make them behave like good little children

Synoptic tools
* Fix alignment of the tool menu in synoptic tools and the profile2 tool. And remove top margin from synoptic tools iframe content
* Fix synoptic tool button alignment

Misc.
* Default the footer color to something less blue

* Add aria-hidden='true' to font awesome icons in portal templates
