+++ portal/portal-impl/impl/pom.xml
@@ -115,17 +115,19 @@
             <artifactId>velocity</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.sakaiproject.velocity</groupId>
-            <artifactId>sakai-velocity-util</artifactId>
-        </dependency>
-        <dependency>
             <groupId>xom</groupId>
             <artifactId>xom</artifactId>
             <version>1.1</version>
         </dependency>
 		<dependency>
+            <groupId>com.googlecode.json-simple</groupId>
+            <artifactId>json-simple</artifactId>
+            <version>1.1.1</version>
+        </dependency>
+        <dependency>
 			<groupId>org.mockito</groupId>
 			<artifactId>mockito-core</artifactId>
+            <version>1.10.19</version>
 			<scope>test</scope>
 		</dependency>
         <dependency>
