+++ portal/portal-impl/impl/src/bundle/sitenav.properties
@@ -30,7 +30,7 @@ sit_othersites = Groups, etc
 sit_coursesites = Courses
 sit_accessibility = go to accessibility information page
 sit_jumpcontent = jump to content
-sit_menutools = tools
+sit_menutools = Tools
 sit_jumptools = jump to tools list
 sit_jumpworksite = jump to site list
 sit_jumpdrawer = open drawer
@@ -76,9 +76,10 @@ sit_favorite_limit_reached=Only the first {0} sites (above) will display in your
 
 sit_noscript_message = Sakai works much better when JavaScript is enabled. Please enable JavaScript in your Browser.
 rs_enterRoleSwapWithRole = Enter {0} View
-rs_exitRoleSwapWithRole = Exit {0} View
+rs_exitRoleSwapWithRole = Exit View
 rs_viewSiteAs = View Site As:
 rs_selectRole = - Select Role -
+rs_menuTooltip = Show View options
 
 bread_separator = >
