+++ portal/portal-render-engine-impl/pack/src/webapp/vm/morpheus/includeSitesNav.vm
@@ -1,6 +1,6 @@
 <nav id="linkNav" role="navigation" aria-labelledby="sitetabs" class="Mrphs-sitesNav" data-max-tools-int="$maxToolsInt" data-max-tools-anchor="${rloader.sit_alltools}">
     <h1 class="skip" tabindex="-1" id="sitetabs">${rloader.sit_worksiteshead}</h1>
-    <div id="show-all-sites" tabindex="-1"><i class="fa fa-angle-double-left"></i><span>${rloader.sit_more}</span></div>
+    <div id="show-all-sites" tabindex="-1"><i class="fa fa-angle-double-left" aria-hidden="true"></i><span>${rloader.sit_more}</span></div>
     <div id="topnav_container">
         <ul class="Mrphs-sitesNav__menu" id="topnav" role="menubar" aria-label="${rloader.sit_worksiteshead}">
 
@@ -9,7 +9,7 @@
         		 #if (${site.isMyWorkspace})
                     <li class="Mrphs-sitesNav__menuitem Mrphs-sitesNav__menuitem--myworkspace #if (${site.isCurrentSite}) is-selected #end">
                         <a class="link-container" href="${site.siteUrl}" title="${rloader.sit_mywor}" role="menuitem">
-                            <i class="fa fa-home"></i>
+                            <i class="fa fa-home" aria-hidden="true"></i>
                             <span class="Mrphs-sitesNav__menuitem--myworkspace-label">${rloader.sit_mywor}</span>
                             <span class="Mrphs-sitesNav__drop" tabindex="-1" data-site-id="${site.siteId}"></span>
                         </a>
