+++ portal/portal-render-engine-impl/pack/src/webapp/vm/morpheus/moresites.vm
@@ -30,7 +30,7 @@
 
             #end ## END of IF (${tabsSites.prefsToolUrl})
 
-            <li class="otherSitesMenuClose"><a href="javascript:void(0);"><i class="fa fa-times"></i></a></li>
+            <li class="otherSitesMenuClose"><a href="javascript:void(0);"><i class="fa fa-times" aria-hidden="true"></i></a></li>
         </ul>
 
         <div id="selectSite">
@@ -39,7 +39,7 @@
                 <li id="tab-btn-courses" class="tab-btn course active" data-tab-target="otherSitesCategorWrap" data-site-type="course"><h2 class="favorites-tab-label">${rloader.sit_coursesites}</h2></li>
                 <li id="tab-btn-others" class="tab-btn other" data-tab-target="otherSitesCategorWrap" data-site-type="other"><h2 class="favorites-tab-label">${rloader.sit_othersites}</h2></li>
                 <li class="organizeFavorites tab-btn" data-tab-target="organizeFavorites">
-                    <h2 class="favorites-tab-label"><span class="favorites-desktop">${rloader.moresite_organize_favorites}</span><span class="favorites-mobile">${rloader.moresite_favorites}</span><span class="favoriteCountAndWarning"> <span class="favoriteCount"></span><span id="favoriteMaxWarningIndicator"> <i class="fa fa-warning"></i></span></span></h2></li>
+                    <h2 class="favorites-tab-label"><span class="favorites-desktop">${rloader.moresite_organize_favorites}</span><span class="favorites-mobile">${rloader.moresite_favorites}</span><span class="favoriteCountAndWarning"> <span class="favoriteCount"></span><span id="favoriteMaxWarningIndicator"> <i class="fa fa-warning" aria-hidden="true"></i></span></span></h2></li>
             </ul>
 
             <div class="tab-pane">
@@ -67,7 +67,7 @@
                             <div class="fav-title #if (${site.isMyWorkspace}) fav-title-myworkspace #end">
                             	<a href="${site.siteUrl}" title="${site.siteTitleNotTruncated}">
                                     #if (${site.isMyWorkspace})
-                                        <i class="fa fa-home"></i><span class="fullTitle">${rloader.sit_mywor}</span>
+                                        <i class="fa fa-home" aria-hidden="true"></i><span class="fullTitle">${rloader.sit_mywor}</span>
                                     #elseif ( ( ${tabDisplayLabel} == 2 ) && ( ${site.shortDescription} ) )
                                         <span class="fullTitle">${site.shortDescription}</span>
                                     #else
@@ -75,7 +75,7 @@
                                     #end
                                 </a>
                             </div>
-                            <a href="#" id="${site.siteId}" class="toolMenus"><i class="fa fa-chevron-down"></i></a>
+                            <a href="#" id="${site.siteId}" class="toolMenus"><i class="fa fa-chevron-down" aria-hidden="true"></i></a>
                         </li>
                     #end
 
@@ -84,9 +84,9 @@
                             #if ($tabsSites.tabsMoreTermsLeftPane.get($termKey).size() > 0)
                                 <div class="fav-sites-term">
                                     #if ( !$termKey || $termKey == "" )
-                                        <h3 class="favorites-term-header"><span class="favorites-select-all-none"></span>${rloader.sit_notermkey}</h3>
+                                        <h3 class="favorites-term-header"><a href='javascript:void(0);' class="favorites-select-all-none"></a>${rloader.sit_notermkey}</h3>
                                     #else
-                                        <h3 class="favorites-term-header"><span class="favorites-select-all-none"></span>$termKey</h3>
+                                        <h3 class="favorites-term-header"><a href='javascript:void(0);' class="favorites-select-all-none"></a>$termKey</h3>
                                     #end
 
                                     <ul class="otherSitesCategorList favoriteSiteList">
@@ -104,7 +104,7 @@
                             #if ($tabsSites.tabsMoreTermsRightPane.get($termKey).size() > 0)
                                 <div class="fav-sites-term">
                                     #if ( $termKey && $termKey != "" )
-                                        <h3 class="favorites-term-header"><span class="favorites-select-all-none"></span>$termKey</h3>
+                                        <h3 class="favorites-term-header"><a href='javascript:void(0);' class="favorites-select-all-none"></a>$termKey</h3>
 
                                         <ul class="otherSitesCategorList favoriteSiteList">
                                             <!-- anchor "my workspace" to the top of the list -->
