+++ portal/portal-render-engine-impl/pack/src/webapp/vm/morpheus/snippets/skipNav-snippet.vm
@@ -15,17 +15,17 @@
                 <span class="accesibility_key">[c]</span>
             </a>
         </li>
-        <li role="menuitem" class="Mrphs-skipNav__menuitem Mrphs-skipNav__menuitem--tools">
-            <a href="#totoolmenu" class="Mrphs-skipNav__link js-toggle-tools-nav" title="${rloader.sit_jumptools}" accesskey="l">
-                <i class="fa fa-cubes tools-icon"></i>
-                ${rloader.sit_menutools}
-                <span class="accesibility_key">[l]</span>
-            </a>
-        </li>
         <li role="menuitem" class="Mrphs-skipNav__menuitem Mrphs-skipNav__menuitem--worksite">
             <a href="#sitetabs" id="more-sites-menu" class="Mrphs-skipNav__link js-toggle-sites-nav" title="${rloader.sit_jumpworksite}" accesskey="w">
-                <i class="fa fa-th all-sites-icon"></i> ${rloader.sit_worksites}
+                <i class="fa fa-th all-sites-icon" aria-hidden="true"></i> ${rloader.sit_worksites}
                 <span class="accesibility_key">[w]</span>
+            </a>
+        </li>
+        <li role="menuitem" class="Mrphs-skipNav__menuitem Mrphs-skipNav__menuitem--tools">
+            <a href="#totoolmenu" class="Mrphs-skipNav__link js-toggle-tools-nav" title="${rloader.sit_jumptools}" accesskey="l">
+                <i class="fa fa-bars tools-icon" aria-hidden="true"></i>
+                ${rloader.sit_menutools}
+                <span class="accesibility_key">[l]</span>
             </a>
         </li>
     </ul>
