+++ reference/library/src/morpheus-master/js/src/sakai.morpheus.more.sites.js
@@ -168,6 +168,8 @@ function showToolMenu(jqObj){
 
     var goToSite = li_template.clone();
 
+    goToSite.addClass('gotosite');
+
     goToSite.find('a')
       .attr('href', portal.portalPath + '/site/' + classId)
       .attr('title', maxToolsText)
@@ -193,7 +195,7 @@ function showToolMenu(jqObj){
 
           // And its icon
           li.find('a span')
-            .addClass('icon-' + item.tools[0].toolId.replace(/\./gi, '-'))
+            .addClass('icon-sakai--' + item.tools[0].toolId.replace(/\./gi, '-'))
             .addClass('otherSiteToolIcon');
 
           if (item.toolpopup) {
