+++ reference/library/src/morpheus-master/sass/base/_responsive.scss
@@ -3,12 +3,34 @@
 body{
 	@media #{$phone}{
 		&.toolsNav--displayed{
-			overflow:hidden;
+			// hide the tool body
+			.#{$namespace}pagebody {
+				#content {
+					display: none;
+				}
+			}
+			// make the banner stick to the top on scroll when the tool menu is visible
+			.#{$namespace}mainHeader {
+				position: absolute;
+				top: -$banner-height;
+
+				.#{$namespace}topHeader {
+					position: relative;
+					padding-top: 0 !important;
+
+					.#{$namespace}headerLogo {
+						position: relative;
+					}
+				}
+			}
+
+			// show the tool menu
 			#toolMenuWrap{
 				width: 100%;
-				top: 6.7em;
-				left: 0px;
 				box-shadow: 0px 0em 1em rgba(0,0,0,0.5);
+				left: 0;
+				z-index: 1001;
+
 				@if $tool-menu-icon-on-left {
 					background-color: $tool-background-color;
 				}
@@ -18,7 +40,7 @@ body{
 				}
 				&:after {
 					bottom: 100%;
-					left: 25%;
+					left: 40px;
 					border: solid transparent;
 					content: " ";
 					height: 0;
@@ -28,7 +50,6 @@ body{
 					border-color: rgba(255, 255, 255, 0);
 					border-bottom-color: $tool-menu-background-color;
 					border-width: 7px;
-					margin-left: -7px;
 					outline: 0;
 				}
 				.Mrphs-toolsNav__menuitem--title{
@@ -96,7 +117,10 @@ body{
 						}
 					}
 				}
+			}
 
+			#portalMask {
+				background: transparent;
 			}
 		}
 		&.sitesNav--displayed{
