+++ reference/library/src/morpheus-master/sass/modules/navigation/_breadCrumbs.scss
@@ -55,7 +55,7 @@
 			font-size: 110%;
 			font-weight: bold;
 			display: inline-block;
-			margin: 1em 0 0.25em 1.25em;
+			margin: 0.75em 0.75em 0.25em 1.25em;
 			border-bottom: 1px solid transparent;
 
 			&:hover{
@@ -86,43 +86,27 @@
 	}
 
 	@media #{$phone}{
-		padding: 1em;
-		background-color: $primary-color;
-		background-image: none;
-		color: $background-color;
-		position: fixed;
-		top: 6.3em;
-		width: 100%;
-		z-index: 2;
-		box-shadow: $divider-shadow;
-		overflow: hidden;
+		background-color: transparent;
+		box-shadow: none;
+		position: absolute;
+		top: $banner-height - 4;
+		height: $banner-height - 4;
+		line-height: $banner-height - 4;
+		width: calc(100% - 80px); // leave space for the Tool menu toggle
 		white-space: nowrap;
-		text-overflow: ellipsis;
-		@include transition( top 0.25s linear 0s );
-		&.moving{
-			top: -22.2em;
-		}
+		left: 80px;
+
 		h1{
-			color: $background-color;
+			color: $breadcrumbs-tool-color;
 		}
-		a{
-			border-bottom: 0 none;
-			color: $background-color;
-			font-size: 0.9em;
-			.#{$namespace}breadcrumb--icon{
-				font-size: 0.95em;
-				vertical-align: 15%;
-			}
-			&:hover{
-				color: $background-color;
-			}
-		}
-		.#{$namespace}hierarchy--siteName a:hover{
+
+		.#{$namespace}hierarchy--siteName a:hover .#{$namespace}hierarchy--siteName-label {
 			text-decoration: underline;
 		}
 		
 		a.#{$namespace}hierarchy--toolName:hover{
-			border-bottom: 1px solid $background-color;
+			color: $breadcrumbs-tool-color;
+			border-bottom: 1px solid $breadcrumbs-hover-color;
 		}
 		
 		.#{$namespace}hierarchy--serviceName, .#{$namespace}hierarchy--serviceNameSeparator{
@@ -131,6 +115,41 @@
 
 		.#{$namespace}breadcrumb--reset-icon{
 			display: none;
+		}
+
+		.#{$namespace}hierarchy--siteName {
+			position: absolute;
+			top: - ($banner-height - 4);
+			height: $banner-height - 4;
+			left: -80px;
+			padding: 0 10px;
+			line-height: $banner-height - 4;
+			overflow: hidden;
+			white-space: nowrap;
+			width: 100%;
+
+			a {
+				color: $sites-nav-menu-item-color;
+				display: inline-block;
+				margin: 0;
+				padding: 0;
+				text-overflow: ellipsis;
+				max-width: 100%;
+				width: auto;
+				overflow: hidden;
+			}
+		}
+
+		.#{$namespace}hierarchy-item {
+			color: $breadcrumbs-tool-color;
+		}
+
+		.#{$namespace}hierarchy--siteNameSeparator {
+			display: none;
+		}
+
+		&.moving {
+			top: -22.2em;
 		}
 	}
