+++ reference/library/src/morpheus-master/sass/modules/swapped-view/_base.scss
@@ -1,18 +1,20 @@
-.#{$namespace}portalWrapper{
-	color: $swapped-view-text-color;
-	&.swapped-view{
-		.#{$namespace}topHeader{
-			background: $swapped-view-primary;
-		}
-		.#{$namespace}siteHierarchy{
-		  background: $background-color-secondary;
-		  color: $background-color;
-		  a {
-		  	color: $background-color;
-		  }
-		}
-		.#{$namespace}mainHeader.is-maximized .siteNavWrap{
-			background: $swapped-view-sites-nav-background;
+@if $swapped-view-enabled {
+	.#{$namespace}portalWrapper{
+		color: $swapped-view-text-color;
+		&.swapped-view{
+			.#{$namespace}topHeader{
+				background: $swapped-view-primary;
+			}
+			.#{$namespace}siteHierarchy{
+				background: $background-color-secondary;
+				color: $background-color;
+				a {
+					color: $background-color;
+				}
+			}
+			.#{$namespace}mainHeader.is-maximized .siteNavWrap{
+				background: $swapped-view-sites-nav-background;
+			}
 		}
 	}
 }
