+++ reference/library/src/morpheus-master/sass/modules/tool/_base.scss
@@ -1,17 +1,11 @@
 
 .#{$namespace}pageColumns--single {
 	.#{$namespace}toolTitleNav__button_container {
-		top: -30px;
-		right: 0;
+		top: -34px;
+		right: 10px;
 		position: absolute;
 		float: none;
 		left: initial;
-	}
-
-  @media #{$tablet} {
-	  .#{$namespace}pageColumns--single {
-	    margin: 0 !important;
-	  }
 	}
 }
 
@@ -36,16 +30,16 @@
 		.#{$namespace}container--toolTitleNav {
 			background-color: $overview-portlet-title-background-color;
 			border-bottom: 1px solid $overview-portlet-border-color;
+			padding: 0.6em;
   	}
   }
 
 	.#{$namespace}toolTitleNav__button_container {
-		display: block;
-		position: relative;
+		position: absolute;
 		float: right;
-		top: -4px;
-		right: -11px;
-		margin-top: -20px;
+		top: 0;
+		right: 4px;
+
 		@media #{$tablet}{
 			font-size: 1.2em;
 		}
@@ -68,7 +62,6 @@
 
 .#{$namespace}toolTitleNav{
 	margin: 0;
-	padding: 0.6em;
 	position: relative;
 	.#{$namespace}toolTitleNav__title{
 		display: none;
@@ -110,64 +103,42 @@
 	}
 
 	.#{$namespace}toolTitleNav__link{
-		color: $tool-nav-link-color;
-		border: 1px solid #e0e0e0;
-		border-radius: 3px;
-		background: $tool-tab-background-color;
+		@include sakai_secondary_button();
+
 		margin-right: 0.25em;
 		padding: 0.4em 0.8em 0.45em 0.8em;
-		text-decoration: none;
 		vertical-align: 0%;
-		&:hover{
-			background: $tool-background-color;
-			color: $swapped-view-text-color;
-			text-decoration: none;
+		line-height: 1em;
+
+		&:before {
+			@extend .fa;
 		}
+
 		@media #{$tablet}{
-			background: $tool-background-color;
-			border: 0px none;
-			color: $primary-color;
 			margin-right: 0em;
 		}
-		/*@media #{$tablet}{
-			background: transparent;
-		}*/
-		//@include icon();
 		&--directurl{
 			@extend .fa;
 			@extend .fa-link;
-			//@extend .fa-lg;
 			/*position: absolute;
 			right: #{$icon-size * 1.5};*/
 		}
 		&--help-popup{
 			@extend .fa;
 			@extend .fa-question;
-			//@extend .fa-lg;
-			/*position: absolute;
-			right: 0;*/
 		}
 		&--edit{
 			@extend .fa;
 			@extend .fa-pencil;
-			//@extend .fa-lg;
-			/*position: absolute;
-			right: #{$icon-size * 3};*/
 		}
 		&--reset{
 			@extend .fa;
 			@extend .fa-share;
-			//@extend .fa-lg;
 			display: none;
 			font-size: 0.9em;
 			float: left;
 		}
 		.#{$namespace}itemTitle{
-			font-family: $font-family;
-			/*font-size: $default-font-size;*/
-			text-transform: uppercase;
-			font-size: 0.8em;
-			letter-spacing: 1px;
 			@media #{$tablet}{
 				display: none;
 			}
@@ -227,10 +198,4 @@
 
 .alertMessage{
 	@extend .bs-callout-danger;
-}
-
-@media #{$tablet} {
-  #content {
-    padding-top: 1em;
-  }
 }
