+++ reference/library/src/morpheus-master/sass/modules/tool/_menu.scss
@@ -1,13 +1,9 @@
 .navIntraTool{
 	margin: -14px -14px 0 -14px;
-	background: $tool-background-color;
 	border-bottom: 1px solid $tool-border-color;
-	padding: 0 170px 0 1em;
+	padding: 0 1em 0 0;
 	//max-width: calc(100% - 170px);
-	@media #{$phone}{
-		padding: 0.3em 0.7em;
-		//max-width: 100%;
-	}
+	font-size: 0;
 	&:after{
 		content: " ";
 		display: block;
@@ -15,54 +11,82 @@
 	}
 	li{
 		display: inline-block;
-		margin: 0 0 0 0;
+		font-size: $default-font-size;
+		margin: 0 -1px 0 0;
 		span{
 			display: inline;
 			a{
 				background: $tool-tab-background-color;
-				border-top: 1px solid $tool-border-color;
-				border-left: 1px solid $tool-border-color;
-				border-right: 1px solid $tool-border-color;
-				color: $tool-nav-link-color;
+				border-top: 1px solid $tool-tab-border-color;
+				border-left: 1px solid $tool-tab-border-color;
+				border-right: 1px solid $tool-tab-border-color;
+				color: $tool-tab-text-color;
+				font-size: $tool-tab-text-size;
+				font-weight: $tool-tab-text-weight;
 				display: block;
-				font-size: 0.8em;
-				margin: 2px 0.25em 0 0;
-				letter-spacing: 1.2px;
-				padding: 6px 10px;
+				margin: 0;
+				padding: 0 0.6em;
+				line-height: $tool-tab-height;
+				height: $tool-tab-height;
 				text-decoration: none;
-				text-transform: uppercase;
-				@media #{$phone}{
-					@include sakai_button( $button-background-color, $button-border-color, $button-text-color, $button-background-secondary-color );
-				}
-				
-				&:hover
-				{
-					color: $swapped-view-text-color;
+				text-transform: $tool-tab-text-transform;
+				letter-spacing: $tool-tab-text-letter-spacing;
+
+				&:hover{
+					color: $tool-tab-hover-text-color;
+					background: $tool-tab-hover-background-color;
 					text-decoration: none;
-				}
+					position: relative;
+
+					&:before{
+						border-top: 4px solid $tool-tab-hover-highlight-color;
+						content: "";
+						display: block;
+						position: absolute;
+						left: -1px;
+						right: -1px;
+						top: -1px;
+					}
+			  }
 			}
 			&.current{
-				background: $tool-background-color;
-				border-top: 1px solid $tool-border-color;
-				border-left: 1px solid $tool-border-color;
-				border-right: 1px solid $tool-border-color;
-				display: block;
-				font-size: 0.8em;
-				letter-spacing: 1.2px;
-				margin: 2px 0.25em 0 0;
-				padding: 7px 10px 6px 10px;
+				color: $tool-tab-active-text-color;
+				font-size: $tool-tab-active-text-size;
+				background-color: $tool-tab-active-background-color;
+				border-top: 1px solid $tool-tab-border-color;
+				border-left: 1px solid $tool-tab-border-color;
+				border-right: 1px solid $tool-tab-border-color;
 				position: relative;
-				text-transform: uppercase;
-				top: 1px;
-				@media #{$phone}{
-					@include sakai_button_color( $button-color-background-color, $button-color-border-color, $button-color-text-color, $button-color-background-secondary-color );
+				margin: 0;
+				padding: 0 0.6em;
+				line-height: $tool-tab-height;
+				height: $tool-tab-height;
+				display: block;
+				margin-bottom: -1px;
+
+				&:before {
+					border-top: 4px solid $tool-tab-active-highlight-color;
+					content: "";
+					display: block;
+					position: absolute;
+					left: -1px;
+					right: -1px;
+					top: -1px;
+				}
+				&:after{
+					border-bottom: 1px solid $tool-tab-active-background-color;
+					content: '';
+					display: block;
+					position: absolute;
+					left: 0;
+					right: 0;
+					bottom: -1px;
 				}
 				a{
-					background: #FFF;
 				    border: 0px none;
 				    font-size: 1em;
-				    margin: 0 0 0 0;
-				    padding: 0 0 0 0;
+				    margin: 0;
+				    padding: 0;
 				    top: 1px;
 				}
 			}
@@ -71,12 +95,12 @@
 			}
 		}
 		a{
-			
+
 			&.current{
 			}
 		}
 		.searchField{
-			
+
 		}
 	}
 }
@@ -108,7 +132,7 @@
 		legend{
 			display: none;
 		}
-		
+
 		fieldset{
 			display:inline-block;
 		}
