+++ reference/library/src/morpheus-master/sass/modules/tool/profile2/_profile2.scss
@@ -18,7 +18,10 @@
 	}
 
 	.navIntraTool {
-	    margin: -2px -14px 0 -14px;
+	    margin: -2px -14px 0 0;
+	    @media #{$phone}{
+	    	margin-top: 10px;
+	    }
 	}
 
 	.Mrphs-pageColumns--single {
