+++ reference/library/src/morpheus-master/sass/modules/tool/synoptic/_synoptic.scss
@@ -1,3 +1,11 @@
+body > .portletBody {
+	margin: 0;
+
+	> .navIntraTool {
+		margin: 0;
+	}
+}
+
 .#{$namespace}sakai-synoptic-announcement, .#{$namespace}sakai-summary-calendar, .#{$namespace}sakai-synoptic-messagecenter, .#{$namespace}sakai-synoptic-chat{
 	iframe{
 		z-index: 0;
@@ -12,4 +20,8 @@
 body > .portletBody > #calendarForm{
 	overflow: hidden;
 	padding: 12px 0 0 4px;
+
+	.navIntraTool {
+		margin: 0;
+	}
 }
